requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
PyYAML>=6.0.1
python-dateutil>=2.8.2
tenacity>=8.2.3
